refactor(qc): consolidate duplicated data loading into data_utils.py (#652)
Extract load_data, generate_synthetic_data, compute_data_hash from 4 training
scripts (classification, lstm, transformer, dqn) into shared data_utils module.
~180 lines of duplication removed.
